requirements.txt
python-dotenv
streamlit
pickle
pandas